# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "description": "Exactly one of ``site_id`` / ``roi`` locates the analysis. ``seed`` makes\nthe Monte Carlo reproducible; a re-run with the same seed is bit-for-bit.",
        "properties": {
          "k_sigma": {
            "default": 2.0,
            "maximum": 5.0,
            "minimum": 0.5,
            "title": "K Sigma",
            "type": "number"
          },
          "method": {
            "default": "mbmp",
            "enum": [
              "mbmp",
              "mbsp"
            ],
            "title": "Method",
            "type": "string"
          },
          "min_area_px": {
            "default": 5,
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Min Area Px",
            "type": "integer"
          },
          "reference_scene_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Scene Id"
          },
          "roi": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBoxIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "source_lonlat": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Lonlat"
          },
          "target_scene_id": {
            "minLength": 1,
            "title": "Target Scene Id",
            "type": "string"
          }
        },
        "required": [
          "target_scene_id"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "BBoxIn": {
        "properties": {
          "east": {
            "title": "East",
            "type": "number"
          },
          "kind": {
            "const": "bbox",
            "default": "bbox",
            "title": "Kind",
            "type": "string"
          },
          "north": {
            "title": "North",
            "type": "number"
          },
          "south": {
            "title": "South",
            "type": "number"
          },
          "west": {
            "title": "West",
            "type": "number"
          }
        },
        "required": [
          "west",
          "south",
          "east",
          "north"
        ],
        "title": "BBoxIn",
        "type": "object"
      },
...
      "DetectionDetailOut": {
        "description": "Full detail: numbers, params, mask + overlay geometry, validation.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "title": "Flags",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "ime_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ime Kg"
          },
          "mask_geojson": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mask Geojson"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "overlay_bounds": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlay Bounds"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "q_kg_h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q Kg H"
          },
          "q_sigma_kg_h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q Sigma Kg H"
          },
          "reference_scene_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Scene Id"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "scene_id": {
            "title": "Scene Id",
            "type": "string"
          },
          "scene_time_utc": {
            "title": "Scene Time Utc",
            "type": "string"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "candidate",
              "accepted",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "u10_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "U10 Ms"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "validation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation"
          },
          "wind_from_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind From Deg"
          },
          "xch4_max_ppb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xch4 Max Ppb"
          }
        },
        "required": [
          "id",
          "site_id",
          "source",
          "status",
          "method",
          "scene_id",
          "scene_time_utc",
          "q_kg_h",
          "q_sigma_kg_h",
          "xch4_max_ppb",
          "u10_ms",
          "wind_from_deg",
          "flags",
          "created_at",
          "updated_at",
          "reference_scene_id",
          "ime_kg",
          "notes",
          "result",
          "params",
          "mask_geojson",
          "overlay_bounds",
          "validation"
        ],
        "title": "DetectionDetailOut",
        "type": "object"
      },
      "DetectionOut": {
        "description": "Summary row for the detection feed (headline numbers only).",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "title": "Flags",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "q_kg_h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q Kg H"
          },
          "q_sigma_kg_h": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q Sigma Kg H"
          },
          "scene_id": {
            "title": "Scene Id",
            "type": "string"
          },
          "scene_time_utc": {
            "title": "Scene Time Utc",
            "type": "string"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "candidate",
              "accepted",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "u10_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "U10 Ms"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "wind_from_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind From Deg"
          },
          "xch4_max_ppb": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xch4 Max Ppb"
          }
        },
        "required": [
          "id",
          "site_id",
          "source",
          "status",
          "method",
          "scene_id",
          "scene_time_utc",
          "q_kg_h",
          "q_sigma_kg_h",
          "xch4_max_ppb",
          "u10_ms",
          "wind_from_deg",
          "flags",
          "created_at",
          "updated_at"
        ],
        "title": "DetectionOut",
        "type": "object"
      },
      "DetectionPatch": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "candidate",
                  "accepted",
                  "rejected"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "DetectionPatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobCreated": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "JobCreated",
        "type": "object"
      },
...
      "SiteIn": {
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/BBoxIn"
          },
          "date_hint_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Hint End"
          },
          "date_hint_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Hint Start"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "name",
          "bbox"
        ],
        "title": "SiteIn",
        "type": "object"
      },
      "SiteOut": {
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/BBoxIn"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "date_hint_end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Hint End"
          },
          "date_hint_start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Hint Start"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "id",
          "name",
          "bbox",
          "date_hint_start",
          "date_hint_end",
          "notes",
          "created_at"
        ],
        "title": "SiteOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenEarth API",
    "version": "2.0.0a0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/methane/analyze": {
      "post": {
        "operationId": "submit_analyze_api_methane_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Analyze",
        "tags": [
          "methane"
        ]
      }
    },
    "/api/methane/detections": {
      "get": {
        "operationId": "list_detections_api_methane_detections_get",
        "parameters": [
          {
            "in": "query",
            "name": "site_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DetectionOut"
                  },
                  "title": "Response List Detections Api Methane Detections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Detections",
        "tags": [
          "methane"
        ]
      }
    },
    "/api/methane/detections/{det_id}": {
      "delete": {
        "operationId": "delete_detection_api_methane_detections__det_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "det_id",
            "required": true,
            "schema": {
              "title": "Det Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Detection",
        "tags": [
          "methane"
        ]
      },
      "get": {
        "operationId": "get_detection_api_methane_detections__det_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "det_id",
            "required": true,
            "schema": {
              "title": "Det Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Detection",
        "tags": [
          "methane"
        ]
      },
      "patch": {
        "operationId": "patch_detection_api_methane_detections__det_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "det_id",
            "required": true,
            "schema": {
              "title": "Det Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectionDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Detection",
        "tags": [
          "methane"
        ]
      }
    },
    "/api/methane/detections/{det_id}/array.npz": {
      "get": {
        "operationId": "detection_array_api_methane_detections__det_id__array_npz_get",
        "parameters": [
          {
            "in": "path",
            "name": "det_id",
            "required": true,
            "schema": {
              "title": "Det Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detection Array",
        "tags": [
          "methane"
        ]
      }
    },
    "/api/methane/detections/{det_id}/overlay.png": {
      "get": {
        "operationId": "detection_overlay_api_methane_detections__det_id__overlay_png_get",
        "parameters": [
          {
            "in": "path",
            "name": "det_id",
            "required": true,
            "schema": {
              "title": "Det Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vmin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vmin"
            }
          },
          {
            "in": "query",
            "name": "vmax",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vmax"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/png": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detection Overlay",
        "tags": [
          "methane"
        ]
      }
    },
    "/api/methane/sites": {
      "get": {
        "operationId": "list_sites_api_methane_sites_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SiteOut"
                  },
                  "title": "Response List Sites Api Methane Sites Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sites",
        "tags": [
          "methane"
        ]
      },
      "post": {
        "operationId": "create_site_api_methane_sites_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Site",
        "tags": [
          "methane"
        ]
      }
    },
    "/api/methane/sites/{site_id}": {
      "delete": {
        "operationId": "delete_site_api_methane_sites__site_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Site",
        "tags": [
          "methane"
        ]
      },
      "patch": {
        "operationId": "patch_site_api_methane_sites__site_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SitePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Site",
        "tags": [
          "methane"
        ]
      }
    },
    "/api/methane/sites/{site_id}/scenes": {
      "get": {
        "operationId": "list_site_scenes_api_methane_sites__site_id__scenes_get",
        "parameters": [
          {
            "in": "path",
            "name": "site_id",
            "required": true,
            "schema": {
              "title": "Site Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "max_cloud",
            "required": false,
            "schema": {
              "default": 80.0,
              "maximum": 100,
              "minimum": 0,
              "title": "Max Cloud",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SceneInfoOut"
                  },
                  "title": "Response List Site Scenes Api Methane Sites  Site Id  Scenes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Site Scenes",
        "tags": [
          "methane"
        ]
      }
    },
...
  }
}